Move Qualifier into separate module.
